working build, runs and at least some commands are working. Good progress!
Also, lots of time spent trial and error with different test frameworks, settled on Mockito and TestNG and got first test working with a test pattern I'm happy with. That will make it easier to validate everything is working as I put all the pieces back together after the major refactor.